+++ app/models/survey/AnswerFormat.java
@@ -1,14 +1,18 @@
 /*******************************************************************************
  *        File: AnswerFormat.java
- *    Revision: 1
+ *    Revision: 2
  * Description: 
  *      Author: Morteza Ansarinia <ansarinia@me.com>
  *  Created on: July 3, 2015
- *     Project: itrc.cset
+ *     Project: itrc.ics
  *   Copyright: See the file "LICENSE" for the full license governing this code.
  *******************************************************************************/
 package models.survey;
 
-public class AnswerFormat {
-
+//TODO make it an enitity to apply more complex answer formats.
+public enum AnswerFormat {
+  STATEMENT,
+  TEXT,
+  MULTIPLE_CHOICE,
+  SCALE
 }